+++ accepton/request.py
@@ -29,13 +29,22 @@ class Request(object):
         self.headers = Headers(client).request_headers()
 
     def perform(self):
+        args = self.__build_request_args()
         response = requests.request(self.request_method,
                                     self.uri.geturl(),
-                                    headers=self.headers,
-                                    json=self.options)
+                                    **args)
         response_body = Response(response.json(object_hook=self.__deserialize))
         return self.__fail_or_return_response_body(response_body,
                                                    response.status_code)
+
+    def __build_request_args(self):
+        options_key = 'params' if self.request_method == 'get' else 'json'
+
+        args = {}
+        args["headers"] = self.headers
+        args[options_key] = self.options
+
+        return args
 
     def __default_options(self):
         return {"environment": "production"}