+++ tests/test_request.py
@@ -1,4 +1,5 @@
 import httpretty
+import json
 import unittest
 from accepton.client import Client
 from accepton.request import Request
@@ -37,3 +38,46 @@ class RequestTest(unittest.TestCase):
         self.request.perform()
         self.assertEqual(httpretty.last_request().headers["user-agent"],
                          self.client.user_agent)
+
+
+class GetRequestTest(unittest.TestCase):
+
+    def setUp(self):
+        self.client = Client(api_key="skey_123")
+        self.params = {"test": "value", "test2": "value2"}
+        self.request = Request(self.client, "get", "/path", self.params)
+        httpretty.enable()
+        httpretty.register_uri(httpretty.GET,
+                               "https://checkout.accepton.com/path",
+                               body='{}')
+
+    def tearDown(self):
+        httpretty.disable()
+        httpretty.reset()
+
+    def test_that_params_are_query_string(self):
+        self.request.perform()
+        self.assertEqual(httpretty.last_request().querystring,
+                         {"test": ["value"], "test2": ["value2"]})
+        self.assertEqual(httpretty.last_request().body, b"")
+
+
+class PostRequestTest(unittest.TestCase):
+
+    def setUp(self):
+        self.client = Client(api_key="skey_123")
+        self.params = {"test": "value", "test2": "value2"}
+        self.request = Request(self.client, "post", "/path", self.params)
+        httpretty.enable()
+        httpretty.register_uri(httpretty.POST,
+                               "https://checkout.accepton.com/path",
+                               body='{}')
+
+    def tearDown(self):
+        httpretty.disable()
+        httpretty.reset()
+
+    def test_that_params_are_json(self):
+        self.request.perform()
+        body = json.loads(httpretty.last_request().body.decode("utf-8"))
+        self.assertEqual(body, {"test": "value", "test2": "value2"})