+++ CHANGELOG.rst
@@ -7,8 +7,8 @@ this scheme are considered to be bugs.
 
 .. _Semantic Versioning 2.0.0: http://semver.org/spec/v2.0.0.html
 
-`Unreleased`_
--------------
+`0.3.0`_ - 2015-11-06
+---------------------
 
 Added
 ~~~~~
@@ -39,5 +39,6 @@ Added
 
 - Initial version.
 
+.. _0.3.0: https://github.com/accepton/accepton-python/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/accepton/accepton-python/compare/v0.1.0...v0.2.0
-.. _Unreleased: https://github.com/accepton/accepton-python/compare/v0.2.0...HEAD
+.. _Unreleased: https://github.com/accepton/accepton-python/compare/v0.3.0...HEAD